+++ Code/WWLib/ffactory.cpp
@@ -169,7 +169,7 @@ void SimpleFileFactoryClass::Append_Sub_Directory(const char *sub_directory) {
   CriticalSectionClass::LockClass lock(Mutex);
 
   // Ensure a trailing semicolon is present, unless the directory list is empty
-  int len = SubDirectory.Get_Length();
+  size_t len = SubDirectory.Get_Length();
   if (len && SubDirectory[len - 1] != ';') {
     SubDirectory += ';';
   }
